requirements.txt
# Core ML / ESM3
esm>=3.0.0
torch>=2.0.0

# ESM2 log-likelihood scoring (HuggingFace)
transformers>=4.40.0

# Anthropic Claude API (NL parser)
anthropic>=0.25.0

# Streamlit UI
streamlit>=1.35.0
stmol>=0.0.9
py3Dmol>=2.0.0

# Bioinformatics
biopython>=1.83

# Data / utils
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
